created are you sure you want to delete modal

+++ client/src/components/Movie.js
@@ -3,10 +3,13 @@ import { Link, useParams, useHistory } from 'react-router-dom';
 
 import axios from 'axios';
 
+import DeleteModal from './DeleteModal';
+
 const Movie = (props) => {
     const { addToFavorites, deleteMovie } = props;
 
     const [movie, setMovie] = useState('');
+    const [showModal, setShowModal] = useState(false)
 
     const { id } = useParams();
     const { push } = useHistory();
@@ -22,6 +25,10 @@ const Movie = (props) => {
     }, [id]);
 
     const handleDeleteClick = () => {
+        setShowModal(true)
+    }
+
+    const deleteRequest = () => {
         axios.delete(`http://localhost:5000/api/movies/${id}`)
         .then(res => {
             deleteMovie(res.data)
@@ -30,6 +37,10 @@ const Movie = (props) => {
         .catch(err => {
             console.log(err)
         })
+    }
+
+    const handleCancel = () => {
+        setShowModal(false)
     }
 
     return(<div className="modal-page col">
@@ -65,6 +76,7 @@ const Movie = (props) => {
                             <Link to={`/movies/edit/${movie.id}`} className="m-2 btn btn-success">Edit</Link>
                             <span className="delete"><input type="button" onClick={handleDeleteClick} className="m-2 btn btn-danger" value="Delete"/></span>
                         </section>
+                        {showModal && <DeleteModal cancelFunction={handleCancel}deleteRequest={deleteRequest}/>}
                     </div>
                 </div>
             </div>
